updated and added policies

+++ app/Http/Controllers/PostController.php
@@ -6,11 +6,23 @@ use App\Models\Post;
 use App\Http\Requests\StorePostRequest;
 use App\Http\Requests\UpdatePostRequest;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Routing\Controllers\Middleware;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\View\View;
 
-class PostController extends Controller
+class PostController extends Controller implements HasMiddleware
 {
+    public static function middleware()
+    {
+        // TODO: Implement middleware() method.
+        return [
+            new Middleware('auth',
+                except: ['index','store']
+            )];
+    }
+
     /**
      * Display a listing of the resource.
      */
@@ -36,7 +48,7 @@ class PostController extends Controller
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request)
+    public function store(Request $request):string
     {
         // validate
         $fields = $request->validate([
@@ -52,7 +64,7 @@ class PostController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(Post $post)
+    public function show(Post $post):string
     {
         //
         return view('posts.show', ['post'=>$post]);
@@ -62,27 +74,44 @@ class PostController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(Post $post)
+    public function edit(Post $post):View
     {
         //
+        Gate::authorize('modify',$post);
+
+        return view('posts.edit', ['post'=>$post]);
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(UpdatePostRequest $request, Post $post)
+    public function update(Request $request, Post $post)
     {
-        //
+        Gate::authorize('modify',$post);
+
+        // validate
+        $fields = $request->validate([
+            'title' => ['required', 'max:225'],
+            'body' => ['required']
+        ]);
+
+        $post->update($fields);
+
+        return redirect()->route('dashboard')->with("success",
+        'Your post was updated');
+
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Post $post)
+    public function destroy(Post $post):string
     {
         //
+        Gate::authorize('modify',$post);
         $post->delete();
-        
+
         return back()->with('delete', 'Your post was deleted');
     }
+
 }
